# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: ESG ProofOps API
  version: 1.0.0
  description: Implementation contract, not a deployed service. Domain v2.0 is authoritative.
servers:
- url: /
paths:
...
  /v1/runs/{run_id}/source-condition-review:
    post:
      summary: source_condition_publish
      operationId: source_condition_publish
      security:
      - sessionCookie: []
      parameters:
      - name: run_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: X-CSRF-Token
        in: header
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SourceConditionEnvelope'
          headers:
            ETag:
              description: Quoted immutable source review revision, e.g. "1"
              schema:
                type: string
            Cache-Control:
              schema:
                type: string
                const: no-store
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden or CSRF failure
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Not found or tenant mismatch
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Conflict / gate blocked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '412':
          description: Stale revision
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: Schema or domain validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Rate/budget limit
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: Dependency unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '413':
          description: Payload or source fragment inventory exceeds the supported bound
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      x-minimum-role: reviewer
      x-idempotency-required: false
      x-rate-limit: 10/min/user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SourceConditionPublish'
      description: Publish review revision 1 for an already committed immutable source
        snapshot. Requires session, allowed Origin and CSRF. Repeats return the existing
        head; changed source inputs return 409. Source-snapshot idempotence requires
        neither Idempotency-Key nor If-Match. No source fact is approved.
    get:
      summary: source_condition_get
      operationId: source_condition_get
      security:
      - sessionCookie: []
      parameters:
      - name: run_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: revision
        in: query
        required: false
        schema:
          anyOf:
          - type: integer
            minimum: 1
          - type: 'null'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SourceConditionEnvelope'
          headers:
            ETag:
              description: Quoted immutable source review revision, e.g. "1"
              schema:
                type: string
            Cache-Control:
              schema:
                type: string
                const: no-store
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden or CSRF failure
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Not found or tenant mismatch
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Conflict / gate blocked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '412':
          description: Stale revision
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: Schema or domain validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Rate/budget limit
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: Dependency unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '413':
          description: Payload or source fragment inventory exceeds the supported bound
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      x-minimum-role: reviewer
      x-idempotency-required: false
      x-rate-limit: 120/min/user
      description: Read the current or requested immutable revision without publication
        or mutation. Missing publication returns 409 SOURCE_REVIEW_NOT_PUBLISHED. Coverage
        remains unknown. The fragment inventory retains overlaps and unassigned fragments;
        more than 5000 entries returns 413 without truncation.
  /v1/runs/{run_id}/source-condition-review/source-view:
    get:
      summary: source_condition_view
      operationId: source_condition_view
      security:
      - sessionCookie: []
      parameters:
      - name: run_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: revision
        in: query
        required: true
        schema:
          type: integer
          minimum: 1
      - name: fragment_id
        in: query
        required: true
        schema:
          type: string
          pattern: ^[0-9a-f]{64}$
      - name: format
        in: query
        required: false
        schema:
          enum:
          - receipt
          - image
          type: string
          default: receipt
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SourceConditionDisplay'
            image/png:
              schema:
                type: string
                format: binary
          headers:
            ETag:
              description: Quoted immutable source review revision, e.g. "1"
              schema:
                type: string
            Cache-Control:
              schema:
                type: string
                const: no-store, private
            X-Content-Type-Options:
              schema:
                type: string
                const: nosniff
            Referrer-Policy:
              schema:
                type: string
                const: no-referrer
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden or CSRF failure
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Not found or tenant mismatch
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Conflict / gate blocked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '412':
          description: Stale revision
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: Schema or domain validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Rate/budget limit
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: Dependency unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '413':
          description: Payload or source fragment inventory exceeds the supported bound
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      x-minimum-role: reviewer
      x-idempotency-required: false
      x-rate-limit: 120/min/user
      description: Resolve a server-inventoried pinned fragment at the required revision.
        format=receipt returns the display receipt and image_url; format=image returns
        the exact original page PNG. This read publishes no records; unavailable or
        changed pinned inputs return 409 SOURCE_REVIEW_INPUT_MISMATCH.
  /v1/runs/{run_id}/source-condition-review/revisions:
    post:
      summary: source_condition_resolve
      operationId: source_condition_resolve
      security:
      - sessionCookie: []
      parameters:
      - name: run_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: X-CSRF-Token
        in: header
        required: true
        schema:
          type: string
      - name: Idempotency-Key
        in: header
        required: true
        schema:
          type: string
          minLength: 16
          maxLength: 128
      - name: If-Match
        in: header
        required: true
        schema:
          type: string
          pattern: ^"[1-9][0-9]*"$
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SourceAnnotationResult'
          headers:
            ETag:
              description: Quoted immutable source review revision, e.g. "1"
              schema:
                type: string
            Cache-Control:
              schema:
                type: string
                const: no-store
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden or CSRF failure
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Not found or tenant mismatch
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Conflict / gate blocked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '412':
          description: Stale revision
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: Schema or domain validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Rate/budget limit
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: Dependency unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      x-minimum-role: reviewer
      x-idempotency-required: true
      x-rate-limit: 10/min/user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SourceAnnotationWrite'
      description: Factual citation and classification upserts only; the service enforces
        at most 16 total upserts. Ownership, conditions, claim_bindings and numeric_receipts
        remain empty; coverage remains unknown.
  /v1/runs/{run_id}/source-condition-review/observations:
    post:
      summary: source_observations_preview
      operationId: source_observations_preview
      security:
      - sessionCookie: []
      parameters:
      - name: run_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: X-CSRF-Token
        in: header
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SourceObservationPreviewResponse'
          headers:
            Cache-Control:
              schema:
                type: string
                const: no-store
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden or CSRF failure
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Not found or tenant mismatch
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Conflict / gate blocked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '412':
          description: Stale revision
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: Schema or domain validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Rate/budget limit
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: Dependency unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '413':
          description: Payload or source fragment inventory exceeds the supported bound
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      x-minimum-role: reviewer
      x-idempotency-required: false
      x-rate-limit: 10/min/user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SourceObservationPreviewRequest'
      description: Read-only original-bound observation candidates at a pinned historical
        source review revision; no revision or artifact mutation and no numeric grade
        approval.
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: __Host-proofops_session
  schemas:
    Error:
      type: object
      additionalProperties: false
      properties:
        error:
          type: object
          additionalProperties: false
          properties:
            code:
              type: string
            message:
              type: string
            request_id:
              type: string
              format: uuid
            retryable: &id003
              type: boolean
            details:
              type: object
              additionalProperties: false
              properties:
                field:
                  type: string
                reason:
                  type: string
                current_revision:
                  type: integer
                  minimum: 1
                retry_after_seconds:
                  type: integer
                  minimum: 0
          required:
          - code
          - message
          - request_id
          - retryable
      required:
      - error
...
    SourceConditionDisplay:
      properties:
        review_revision:
          type: integer
          minimum: 1.0
        review_sha256:
          type: string
          pattern: ^[0-9a-f]{64}$
        source_snapshot_sha256:
          type: string
          pattern: ^[0-9a-f]{64}$
        fragment_id:
          type: string
          pattern: ^[0-9a-f]{64}$
        display:
          additionalProperties: true
          type: object
        image_url:
          type: string
      additionalProperties: false
      type: object
      required:
      - review_revision
      - review_sha256
      - source_snapshot_sha256
      - fragment_id
      - display
      - image_url
    SourceConditionEnvelope:
      properties:
        review:
          $ref: '#/components/schemas/SourceConditionRevision'
        fragments:
          items:
            $ref: '#/components/schemas/SourceConditionFragment'
          type: array
        issues:
          items:
            $ref: '#/components/schemas/QualityIssue'
          type: array
        coverage_status:
          type: string
          const: unknown
      additionalProperties: false
      type: object
      required:
      - review
      - fragments
      - issues
      - coverage_status
...
    SourceConditionPublish:
      additionalProperties: false
      properties: {}
      type: object
...
    SourceAnnotationResult:
      additionalProperties: false
      properties:
        review:
          $ref: '#/components/schemas/SourceConditionRevision'
        numeric_receipts:
          items:
            additionalProperties: true
            type: object
          maxItems: 0
          type: array
        coverage_status:
          const: unknown
          type: string
      required:
      - review
      - numeric_receipts
      - coverage_status
      type: object
...